+++ homework1/backend/app_init.py
