+++ homework1/backend/route/vm_route.py
@@ -1,16 +1,19 @@
 from flask import Blueprint, request, jsonify, render_template
 from app_svc import VendingMachineManager
-from database.database import VendingMachine
+from database.vending_machine import VendingMachine
 
 vm_controller = Blueprint('vm_controller', __name__)
+
 
 @vm_controller.route('/')
 def index():
     return render_template('base.html')
 
+
 @vm_controller.route('/machines')
 def machines():
     return render_template('machines.html')
+
 
 @vm_controller.route('/add_machine', methods=['POST'])
 def add_machine():
@@ -19,6 +22,8 @@ def add_machine():
     manager = VendingMachineManager()
     manager.create_machine(name, location)
     return jsonify(success=True, message="Vending machine created successfully")
+
+
 @vm_controller.route('/get_machine/<int:id>', methods=['GET'])
 def get_machine(id):
     manager = VendingMachineManager()
@@ -27,6 +32,8 @@ def get_machine(id):
         return jsonify(success=True, machine=machine.to_dict())
     else:
         return jsonify(success=False, message="Vending machine not found"), 404
+
+
 @vm_controller.route('/update_machine/<int:id>', methods=['PUT'])
 def update_machine(id):
     manager = VendingMachineManager()
@@ -38,6 +45,8 @@ def update_machine(id):
         return jsonify(success=True, message="Vending machine updated successfully")
     else:
         return jsonify(success=False, message="Vending machine not found"), 404
+
+
 @vm_controller.route('/delete_machine/<int:id>', methods=['DELETE'])
 def delete_machine(id):
     manager = VendingMachineManager()
@@ -47,6 +56,7 @@ def delete_machine(id):
         return jsonify(success=True, message="Vending machine deleted successfully")
     else:
         return jsonify(success=False, message="Vending machine not found"), 404
+
 
 @vm_controller.route('/all_machines', methods=['GET'])
 def get_all_machines():
